safari-600.7-branch/LayoutTests/inspector-obsolete/http-tests/network/network-sidebar-width.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>
function test()
{
    InspectorTest.makeSimpleXHR("GET", "resources/resource.php", true, step2);

    var oldSetting;
    function step2()
    {
        var panel = WebInspector.panel("network");
        var setting = WebInspector.settings[panel._sidebarWidthSettingName()];
        var request = WebInspector.panel("network").requests[WebInspector.panel("network").requests.length - 1];

        oldSetting = setting.get();
        panel.show();

        setting.set(undefined);
        panel._showRequest(request);
        InspectorTest.addResult("Sidebar width: " + panel.sidebarElement.offsetWidth);
        panel._toggleGridMode();

        setting.set(242);
        panel._showRequest(request);
        InspectorTest.addResult("Sidebar width: " + panel.sidebarElement.offsetWidth);
        panel._closeVisibleRequest(request);

        setting.set(oldSetting);

        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onload="runTest()">
<p>Tests network panel sidebar width with and without preference set.</p>
<a href="https://bugs.webkit.org/show_bug.cgi?id=65256">Bug 65256</a>
</body>
</html>
